% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselector.R
\name{fselector}
\alias{fselector}
\title{Feature Selection for Monotonic Classification}
\usage{
fselector(traindata, trainlabels, k, beta, nselected)
}
\arguments{
\item{traindata}{Training data of numeric type without labels.}

\item{trainlabels}{A vector of numeric tags for each instance of training data.}

\item{k}{positive constant for logit function. If large, fuzzy ordinal set is understood as slightly
larger. If small, is understood as significantly larger.}

\item{beta}{Regulation param for relative importance of MI between features and decision.}

\item{nselected}{Number of features to select.}
}
\value{
nselected most important features.
}
\description{
Selects the N most relevant features from ordinal and monotonic data, based on mRMR criterion.
}
\examples{
data("train_balance-scale")
trainlabels<-dattrain[,ncol(dattrain)]
traindata=dattrain[,-ncol(dattrain)]
selected<-fselector(traindata,trainlabels,2,2,2)


}
