% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iselector.R
\name{iselector}
\alias{iselector}
\title{Instance Selection for Monotonic Classification}
\usage{
iselector(traindata, trainlabels, candidates, collisions, kEdition)
}
\arguments{
\item{traindata}{Training data of numeric type without labels.}

\item{trainlabels}{A vector of numeric tags for each instance of training data.}

\item{candidates}{Rate of the best candidates to be selected.}

\item{collisions}{Minimal rate of collisions permitted to stop the removal process.}

\item{kEdition}{Maximum number of nearest neighborgs to consider.}
}
\value{
A reduced dataset with the selected instances and its labels.
}
\description{
Selects the N most relevant instances from ordinal and monotonic dataset, using a three-steps algorithm.
}
\examples{
dattrain<-read.table("train_balance-scale.0", sep=" ")
trainlabels<-dattrain[,ncol(dattrain)]
traindata=dattrain[,-ncol(dattrain)]
selected<-iselector(traindata,trainlabels,0.01,0.01,5)


}
