% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdlor.R
\docType{class}
\name{kdlorModel-class}
\alias{kdlorModel-class}
\title{Class to model a KDLOR object that is created in the trainning of the model and used in the prediction.}
\description{
Class to model a KDLOR object that is created in the trainning of the model and used in the prediction.
}
