% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdlor.R
\name{kdlorpredict}
\alias{kdlorpredict}
\title{Predicts KDLOR model outputs for test data}
\usage{
kdlorpredict(fittedmodel, trainData, testData)
}
\arguments{
\item{fittedmodel}{fitted model of class kdlorModel obtained with kdlortrain(...)}

\item{trainData}{Data used to previously train the model. Tags should not be provided in traindata.}

\item{testData}{Test data without labels.}
}
\value{
A list of two elements containing the predicted labels for each instances and the projected values.
}
\description{
Predicts the test data labels using a Kernel Discriminant Learning for Ordinal Regression fitted model
}
\examples{
testdata<-read.table("test_balance-scale.0", sep=" ")
testlabels<-testdata[,ncol(testdata)]
testdata<-testdata[,-ncol(testdata)]
pred<-kdlorpredict(myfit,traindata,testdata)


}
