% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_libsvm}
\alias{load_libsvm}
\title{Function to look for libsvm-weights binaries on the system}
\usage{
load_libsvm()
}
\value{
the object containing the library bin, if found. Else throws an error.
}
\description{
The user must install libsvm-weights as explained in the doc. Then this function is used to find the compiled library
on the system.
}
