% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom.R
\name{pomfit}
\alias{pomfit}
\title{Train a proportional odd model for ordinal regression}
\usage{
pomfit(train, trainLabels, linkfunction = "logistic")
}
\arguments{
\item{train}{Training data of numeric type without labels.}

\item{trainLabels}{Tags for each instance of training data. Must be factors.}

\item{linkfunction}{link function to be used in the ordinal logistic regression fit. Possible functions are: logistic','probit','loglog','cloglog' or 'cauchit'.}
}
\value{
the fitted model.
}
\description{
train data must be the training data without labels. Labels should be provided in trainLabels.
}
\examples{
data("train_balance-scale")
fit<-pomfit(dattrain[,-ncol(dattrain)],as.factor(dattrain[,ncol(dattrain)]),"logistic")

}
