% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscala.R
\docType{data}
\name{rscalaBridgeName}
\alias{rscalaBridgeName}
\title{rscala bridge name}
\format{An object of class \code{character} of length 1.}
\usage{
rscalaBridgeName
}
\value{
The Name of the bridge between R and Scala. By default is 's'
}
\description{
Object from rscala package, it returns the name of the bridge between r and Scala
}
\keyword{datasets}
