% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscala.R
\docType{data}
\name{rscalaBridgeOwner}
\alias{rscalaBridgeOwner}
\title{rscala bridge owner}
\format{An object of class \code{logical} of length 1.}
\usage{
rscalaBridgeOwner
}
\value{
TRUE if embeded scala bridge in the package has owner, else FALSE
}
\description{
Object from rscala package, is a boolean pointing if the embeded scala bridge in the package has owner
}
\keyword{datasets}
