% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscala.R
\name{s}
\alias{s}
\title{Scala interpreter instance}
\usage{
s(...)
}
\arguments{
\item{...}{scala code snipped of function to call from R}
}
\value{
The avaiable Scala instance
}
\description{
Scala interpreter instance named 's' by default. Name can be changed using 'assign.name'
}
