% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmop.R
\name{svmopfit}
\alias{svmopfit}
\title{trains the model for the SVMOP method with trainning data and given parameters
train data must be in svmlib format}
\usage{
svmopfit(train, trainLabels, param, weights = TRUE)
}
\arguments{
\item{x}{A number.}

\item{y}{A vector of tags.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
trains the model for the SVMOP method with trainning data and given parameters
train data must be in svmlib format
}
\examples{
add(1, 1)
add(10, 1)

}
