% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wknnor.R
\name{wknnor}
\alias{wknnor}
\title{K-nearest neighbors for ordinal and monotonic data}
\usage{
wknnor(traindata, trainlabels, testdata, k, q, kerneltype, monotonicity)
}
\arguments{
\item{traindata}{Training data of numeric type without labels.}

\item{trainlabels}{A vector of numeric tags for each instance of training data.}

\item{testdata}{Test data of numeric type.}

\item{k}{number of neighbors to use.}

\item{q}{Minkowski distance param. Use q=1 for Manhattan distance and q=2  for Euclidean distance.}

\item{kerneltype}{Kernel used to compute neighbors weights. Avaiable kernels are: rectangular, triangular,epanechnikov,
biweight, triweight, cosine, gauss and inversion.}

\item{monotonicity}{Boolean param specifying whether data is monotone or not.}
}
\value{
Predicted labels for test data.
}
\description{
Predicts labels for test data with ordinal nature or monotonic constraints using an adaptation
of wknn.
}
\examples{
data("train_balance-scale")
traindata=dattrain[,-ncol(dattrain)]
trainlabels=dattrain[,ncol(dattrain)]
data("test_balance-scale")
testlabels<-dattest[,ncol(dattest)]
testdata<-dattest[,-ncol(dattest)]
predictions<-wknnor(traindata,trainlabels,testdata,5,2,"rectangular",FALSE)

}
