% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pom.R
\name{pompredict}
\alias{pompredict}
\title{Predict over the new data instances using the trained model.}
\usage{
pompredict(model, test)
}
\arguments{
\item{model}{A trained POM model.}

\item{test}{Numeric test data without labels.}
}
\value{
A list containing at the first position the projected values per instance per class and at the second position the predicted label for the values.
}
\description{
Predict over the new data instances using the trained model.
}
\examples{
dattrain<-read.csv("train_balance-scale.0", sep=" ")
fit<-pomfit(dattrain[,-ncol(dattrain)],as.factor(dattrain[,ncol(dattrain)]),"logistic")
dattest<-read.csv("test_balance-scale.0", sep=" ")
predictions<-pompredict(fit,dattest[,-ncol(dattest)])
projections<-predictions[[1]]
predictedLabels<-predictions[[2]]

}
