% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmop.R
\name{svmopredict}
\alias{svmopredict}
\title{Predict over the new data instances using the trained models}
\usage{
svmopredict(models, test)
}
\arguments{
\item{models}{A matrix of 1xN trained SVM models. Where N denotes the number of classes of the problem minus one.}

\item{test}{Numeric test data without labels.}
}
\value{
A list containing the projected values per instance per class and the predicted values (the maximum probability for each data instance).
}
\description{
Predict over the new data instances using the trained models
}
\examples{
dattrain<-read.table("train_balance-scale.0", sep=" ")
modelstrain<-svmofit(dattrain[,-ncol(dattrain)],dattrain[,ncol(dattrain)],TRUE,1,1)
dattest<-read.table("test_balance-scale.0", sep=" ")
predictions<-svmopredict(modelstrain,dattest[,-ncol(dattest)])

}
